Tried to fix the login issue -removed the serverHandler Thread -Implemented closing the sockets -I think there's an issue with the scanner reading UTF
